move: domain package 중 MenuJpaRepository 을 infra/repository로 이동
