+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.application;
 import kitchenpos.menus.application.dto.MenuCreateRequest;
 import kitchenpos.menus.application.dto.MenuPriceChangeRequest;
 import kitchenpos.menus.domain.*;
+import kitchenpos.menus.tobe.domain.MenuRepository;
 import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.PurgomalumClient;
 import org.springframework.stereotype.Service;
