+++ src/test/java/kitchenpos/menus/application/InMemoryMenuRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.application;
 
 import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.MenuRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
