+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -6,7 +6,7 @@ import kitchenpos.menus.application.dto.MenuCreateRequest;
 import kitchenpos.menus.application.dto.MenuPriceChangeRequest;
 import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.domain.MenuGroupRepository;
-import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.MenuRepository;
 import kitchenpos.menus.tobe.domain.Product;
 import kitchenpos.clients.FakePurgomalumClient;
 import kitchenpos.products.application.InMemoryProductRepository;
